Prevent regex state explosion in iso9660 signature.
Regex matches with huge repetition counts seem to trigger a state
explosion in the regex engine which can lead to e.g., stack overflows
and also seems to slow down performance significantly.

This patch works around this by not restricting the repetition count.
This should lead to strictly more false positives, but at least makes
this signature useful and not disruptive.

Relates to #3622.