Management framework: add metrics port auto-enumeration
This adds the same logic for automatically assigning ports as already in place
for Broker, just for the separately configurable range of metrics ports.

It also moves the check for whether the user has actually assigned ports when
auto-assignment is disabled into the already existing config validation
function; doing so allows us to throw out some redundant code.