fix: set SignatureFile.full_path if invalid
This sets SignatureFile::full_path member variable to the full path if
SignatureFile::file member is relative and full_path is invalid. This
uses the SignatureFile::load_location to derive the full path from the
file path relative to the load_location.

This is needed when the full_path is passed in to either HookLoadFile
and/or HookLoadFileExtended.