Raise configure error message for unsupported archives
Github lets users download archives of our repos. These do not contain
the necessary submodules. We regularly encounter users who stumble
across this.

We already do have an error message that is raised when a non-recursive
git checkout was done. This commit adds an error message for a non-git
download that does not contain the necessary files.