cirrus: Unset CCACHE_BASEDIR for asan/coverage build
When CCACHE_BASEDIR is set, ccache will rewrite absolute paths to
relative paths in order to allow compilation in different source
directories. We do not use this feature on Cirrus (the checkout
is always in /zeek) and not using absolute paths appears to fix
the coverage build.

We could consider removing the global CCACHE_BASEDIR, but it'd
ccache bust every other task.