cirrus: Unset CCACHE_BASEDIR for asan/coverage build
When CCACHE_BASEDIR is set, ccache will rewrite absolute paths to
relative paths in order to allow compilation in different source
directories. We do not need this feature on Cirrus (the checkout
is always in /zeek) and using absolute paths avoids
confusion/normalization needs for the gcov -p results.

We could consider removing the global CCACHE_BASEDIR, but it'd
bust the ccache of every other task, too.