Management framework: add auto-enumeration of metrics ports
This is quite redundant with the enumeration for Broker ports,
unfortunately. But the logic is subtly different (all nodes obtain a telemetry
port, while not all nodes require a Broker port, for example) and I found more
functional code that abstracts the field names etc less readable in the end.

The logic is for the two remains the same: from a start point, ports get
enumerated sequentially that aren't otherwise taken.

I'm changing the ordering logic slightly since it seems more intuitive to keep
sequential ports on a given instance instead of striping across them.