cluster: Introduce ZeroMQ backend
The ZeroMQ cluster backend uses a central XPUB/XSUB proxy/broker for
Pub/Sub functionality. Logging functionality is realized by logger
nodes that listen on PULL sockets for log writes and all other nodes
connecting to each of the PULL sockets with PUSH sockets.

This requires the cppzmq-dev and libzmq3-dev on Ubuntu 24.04 for
building.