script_opt: Do not include ZBody.h in ZAM/Validate.cc
Touching broker/Manager.h or logging/Manager.h or anything that's
included in ZBody.h results in a rebuild of a bunch of files in ZAM.
Particularly, ZAM/Validate.cc takes much too long to recompile for
me (45+ seconds) and apparently that's mostly by accident.

Clean up the ZAM/Validate.cc includes to prevent this.