[WIP] Make statement lists have a scope
Note that this doesn't actually properly execute shadowing, it simply
makes it not error. And introduces a bunch of other problems.

The reason this doesn't shadow variables is that locals are created at
parse time (hence why a scope is needed in both `Stmt.cc` and `parse.y`
for the statement lists), but locals are used at execution time
(obviously). This means that executing the local will just grab the
outer scope one every time - the inner scope one goes to die.

So the next step *could* be to move locals to getting created at
execution time, which seems like it's what should happen. But that may
be a lot of work.

Optionally, we can alse just add a scope counter and "append" that in
some way to the variable's name. That seems pretty easy.

Not that shadowing is that important to do anyway, I just wanted to try.