misc/stats: Add zeek_net_time_seconds
Adding a metric for the network time value itself should make it
possible to observe it stopping or growing slowly as compared to
realtime when Zeek isn't able to keep up.

Also, modify telemetry/log.zeek to include misc/stats and log at
a higher frequency with a more interesting pcap.