Prevent non-Modbus on port 502 to be reported as Modbus
This commit prevents most non-Modbus TCP traffic on port 502 to be
reported as Modbus in conn.log as well as in modbus.log.
To do so, we have introduced a new function  in the Modbus flow
that checks that some specific fields of the (supposedly) Modbus
header are compatible with values specified in the specs.

The function is then used in guards around the different callback
functions of the parser, to prevent the generation of incorrect
Modbus events.

Closes #3962