Make empty container expansion more intuitive
Closes #3933

There are a couple ways to get a list, sometimes with `{}`, sometimes
with `[]`. You can use these to append to sets and vectors. But,
sometimes it's useful to expand the elements of a list, like if you pass
in a global when appending a set. So, those elements are evaluated and
possibly change the list.

This can, however, end up with an empty list that seems like it should
be impactful for construction. It seems useful to say that an empty list
is intentional, since the initial check is not able to catch those
cases. So, keep that in the list if it's empty.