DPD: log analyzers that have confirmed
This switches the DPD logic to always log analyzers that raised a
protocol confirmation.

The logic is that, once a protocol has been confirmed - and thus there
probably is log output - it does not make sense to later remove it from
the log. It does make sense to somehow flag it as failed - but that
seems like a secondary step.

While doing this, I noticed that the IRC analyzer always confirms
connections, pretty much without regard what happened in it. This commit
also changes the logic of the IRC analyzer to check for valid commands
before confirming.